+++ lib/src/enums/buttons_alignment.dart
@@ -1,0 +1,6 @@
+enum ButtonsAlignment {
+  topLeft,
+  topRight,
+  bottomLeft,
+  bottomRight,
+}